Issue #6 solved Area, delay, explicitDur, action and transition duration now supports real numbers.

+++ js/src/player/Player.js
@@ -281,7 +281,7 @@ Player.prototype.setProperty = function (name, value) {
 
                             case "explicitDur": {
                                     var buffer = value.split("s");
-                                    this.explicitDur = parseInt(buffer[0]);
+                                    this.explicitDur = parseFloat(buffer[0]);
                                     break;
                             }
                             case "baseDeviceRegion":
